Document IDs must be Long, not Int.

+++ worker/src/main/scala/ClusterTypes.scala
@@ -16,7 +16,7 @@ import scala.collection.mutable.ArrayBuffer
 
 object ClusterTypes {
 
-  type DocumentID = Int
+  type DocumentID = Long
   type TermWeight = Float
   type TermID = Int
