+++ worker/src/test/scala/persistence/NodeWriterSpec.scala
@@ -100,7 +100,7 @@ class NodeWriterSpec extends DbSpecification {
     "insert document into node_document table" in new DbTestContext {
       val documentSetId = insertDocumentSet("NodeWriterSpec")
       val documentIds = for (i <- 1 to 5) yield 
-        insertDocument(documentSetId, "title", "documentCloudId").toInt
+        insertDocument(documentSetId, "title", "documentCloudId")
       val idSet = Set(documentIds: _*)
       
       val node = new DocTreeNode(idSet)
